requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
loguru==0.7.2
peewee==3.17.0
pyTelegramBotAPI==4.9.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot-calendar==1.0.5
requests==2.31.0
six==1.16.0
urllib3==2.0.6
